summary: only filter out framing messages that match the test name
Previously the summary would filter out all messages which include
a '--- FAIL' or '--- SKIP'. With this change only the framing
message for the test case will be removed.

This change is in preparation for working around the test2json
output misattribution bug, which will result in the summary
including the output for all subtests in some cases.

After that bug is fixed, and the workaround is removed, it may still
be valuable to be more strict about this filtering.