Use an ID to index TestCase output
So that when a test case is encountered multiple times, either because
-count=n was used, or because ScanTestOutput is being passed multiple instances
of test runs, a failure or skipped test displays only its own output,
not that of all tests with the same name.

Also move addEvent to Package, it is more appropriate as a Package
method, and make Package methods use a pointer receiver. Many of them
update the Package.