Add rerun for --watch mode
When watching for file changes, pressing the 'r' key will re-run tests for the last
package that was run.

Also, mostly incidentally because of the terminal mode, removes echo'ing of characters
pressed while in watch mode. The newline character is handled as a special case
and printed to stdout so that it is still possible to clear part of the screen while
in watch mode.

Only works on unix, and has only been tested on linux.