watch: properly reset terminal after idle timeout
Previous the terminal would not be reset because the defer was in the
redo handler goroutine. This goroutine accepts a context, but because it
blocks on reading from stdin, it does not exit when the context is
cancelled.

To fix the problem, the defer is moved into the main goroutine.