watch: move lastPath storage to handler
This commit fixes a bug where attempting to redo after a debug run would run tests in the wrong
package.

The problem was that debug would create a test binary in the working directory. The creation of
that file would trigger an event. The event wouldn't be handled, but it still caused
prevPath to be updated.

By moving the storage of the lastPath to handler it removes the need for the Save method on
redoHandler, and keeps the tracking of the previous run details (last time.Time) together.