Try out a new pattern for large table tests
Once a table test grows past a certain point it becomes very difficult
to work with. A failure in one case requires scrolling around attempting
to find the case, because the failure message only includes the line in
the test function, not the testCase.

Instead of a slice or map of test cases, using this runCase function to
add a t.Log will give the line number of the testCase as well as the
line number in the test function. This makes it very easy to jump to
both of the relevant lines.