slowest: skip root tests when subtests are slow
Initially I assumed that the root test would include all the time of the
subtests, but that is not always the case. With t.Parallel it appears
that root tests can report 0 elapsed time, while subtests can be very
slow.

To fix this problem, if slowest receives any subtest as one of the slow
tests it will skip the root test. This is necessary because adding a
skip statement to a subtest is substantially more complicated than
adding one to the root test. It's much harder to find the correct place
in the AST to add the if statement.