format: handle 'testing: warning: no tests to run' event
An output event with this text is sent when 'go test' runs a package that has test files
but no tests.

Update two formts that were printing this message in a way that was confusing. Now the standard-quiet format
properly matches 'go test' format, and the testname format omits the text. Since gotestsum includes a count
of tests run in the summary line, this message doesn't seem like it provides much value.