Try a character from dingbats for empty packages
The other two hivis characters are in the dingbat unicode block, using a
code from this block from the third is nice because it increases the
chances that a font supports all 3 characters.

This character is also a bit less bright, which seems nice for skipped
tests, because the focus should be more on the failures.