[ci-matrix] Distribute unknown packages between shards.
Currently all new packages which are not in timing files are assigned to
a single bucket with shortest duration. This happened because duration
of "unknown" packages is set to 0 but in reality it might be longer.
Also empty packages with no tests have 0s duration but there is little
overhead to run those tests.

To avoid all such packages adding into a single bucket, change how those
are distributed between shards. If package run duration is lower than
threshold - assign it by using hash function instead of minimal bucket.