Cleanup some tests
Fix some names to match convention

Remove unnecessary defer, now handled by t.Cleanup since go1.14

Change defaultNoColor to be called when flags are setup so taht the value of
NoColor can be patched. Also patch the value so that when it runs outside of
CI without a terminal the test does not fail.