Add testdox format
This format shows test names as full sentences, which is useful both for
autogenerating documentation ("dox"), and for encouraging programmers to
write their test names as sentences describing the behaviour of the unit
under test.

The specific method of formatting test names as sentences is that used
by https://github.com/bitfield/gotestdox.