Restore default format
Only use github-actions format if format is set to testname. The current default
pkgname is much more compact. Maybe we could add a github-actions-pkgname format
in the future that folds the entire package.

For now it seems safe to use github-actions instead of pkgname if the GITHUB_ACTIONS
environment variable is set.  If someone really wants the output expanded by default
they can unset or overwrite that env var for the 'gotestsum' process.