filewatcher: Silence warning from os.Stat failures
Although unlikely, it is possible for os.Stat to fail in handleDirCreated. The
two most likely causes are a symlink to a target that doesn't exist or a race
condition with a process that deletes the path.

In any case, it isn't actionable by the user and causes no harm to ignore the
error.