tolerate `run` without `pass/failed` for benchmarks
Something in Go 1.20 changed so that there is a `run` event for benchmarks. As
before, there is no `pass` or `failed`. This triggers the "test is running and
thus must have failed" logic in gotestsum, which causes it to report the
benchmark as failed.

As this is the behavior of Go (whether it's correct or not...), gotestsum
should accept such output. To reduce the risk that genuine problems go
undetected, such incomplete benchmarks get reported when a panic was
detected (the original motivation for this workaround).